feat(highnotes): add method to merge highnotes in horizontal line
To parse lines which starts with a hignote and contains highnotes it is
required to merge them to one line. This is done by detecting the most
left highnote which are interested as line start and all notes with a
special distance to this position are threaded as online high notes.